[Agent] Refactor ProcessingCommandState helpers
